py-distance-class

+++ app/main.py
@@ -3,17 +3,17 @@ from __future__ import annotations
 
 class Distance:
 
-    def __init__(self, km: int) -> None:
+    def __init__(self, km: float) -> None:
         self.km = km
 
-    def __add__(self, other: Distance | int) -> Distance:
+    def __add__(self, other: Distance | float) -> Distance:
         if isinstance(other, Distance):
             return Distance(
                 km=self.km + other.km
             )
         return Distance(km=self.km + other)
 
-    def __iadd__(self, other: Distance | int) -> Distance:
+    def __iadd__(self, other: Distance | float) -> Distance:
         if isinstance(other, Distance):
             self.km += other.km
             return self
@@ -26,35 +26,35 @@ class Distance:
     def __repr__(self) -> str:
         return f"Distance(km={self.km})"
 
-    def __mul__(self, other: int) -> Distance:
+    def __mul__(self, other: float) -> Distance:
         return Distance(km=self.km * other)
 
-    def __truediv__(self, other: int) -> Distance:
+    def __truediv__(self, other: float) -> Distance:
         return Distance(
             km=round(self.km / other, 2)
         )
 
-    def __eq__(self, other: Distance | int) -> bool:
+    def __eq__(self, other: Distance | float) -> bool:
         if isinstance(other, Distance):
             return self.km == other.km
         return self.km == other
 
-    def __gt__(self, other: Distance | int) -> bool:
+    def __gt__(self, other: Distance | float) -> bool:
         if isinstance(other, Distance):
             return self.km > other.km
         return self.km > other
 
-    def __lt__(self, other: Distance | int) -> bool:
+    def __lt__(self, other: Distance | float) -> bool:
         if isinstance(other, Distance):
             return self.km < other.km
         return self.km < other
 
-    def __ge__(self, other: Distance | int) -> bool:
+    def __ge__(self, other: Distance | float) -> bool:
         if isinstance(other, Distance):
             return self.km >= other.km
         return self.km >= other
 
-    def __le__(self, other: Distance | int) -> bool:
+    def __le__(self, other: Distance | float) -> bool:
         if isinstance(other, Distance):
             return self.km <= other.km
         return self.km <= other
